cleanup

+++ cute_kernels/jit.py
@@ -70,7 +70,8 @@ def _get_cpp_function(function_name: str, source_files: list[str], build_directo
             verbose=True,
         )
 
-        rmtree(build_directory, ignore_errors=True)
+        if _WORLD_SIZE > 1:
+            rmtree(build_directory, ignore_errors=True)
 
     return getattr(module, function_name)
